project2-rag-system/requirements.txt
# 项目2依赖（简化版）
streamlit>=1.31.0
openai>=1.0.0
python-dotenv>=1.0.0
numpy>=1.0.0
pytest>=7.4.0
